Rename CreatedRecipesSection
